cli: plugins show -> self show plugins
